Pre-rip script improvements

+++ wazzuf-pre-ripper.sh
@@ -15,22 +15,53 @@ checkandsource_wazzuf_files
 check_mplayer
 check_ffmpeg
 
-	#--------------------------------------------------------
-	# Run cropdetect to establish the correct crop rectangle
-	#--------------------------------------------------------
-	CROP_FRAMES=100
-	cropdetect(){
-		echo -n "Running Crop Detection... "
+## pre-rip functions
+
+# Run cropdetect to establish the correct crop rectangle
+CROP_FRAMES=100
+cropdetect(){
+	echo -n "Running Crop Detection... "
+	if [[ "$ISO_FILE" =~ \ |\' ]]
+	then
+		echo -ne "\n *************************************\n"
+		echo " Warning: Path/filename to iso file $ISO_FILE must not contain spaces for croptetection.."
+		echo -ne " *************************************\n"
+	else
 		mplayer ${1} -vf cropdetect -nosound -vo null -frames $CROP_FRAMES -sstep 1 -nocache &> /tmp/cropdetect.out
 		wait
 		crop=$(cat /tmp/cropdetect.out | tr '\r' '\n' | grep 'vf crop\=' | tail -1 | awk -F\= '{print $2}' | awk -F\) '{print $1}')
-		echo -ne "\nEstablished crop factor = ${crop}\n"
+		echo -ne "\nVIDEO_CROP=\"${crop}\"\n"
 		rm /tmp/cropdetect.out
-	}
+	fi
+}
+
+	## 'decode' avconv output file
+avconv-decode(){
+	while read line
+	do
+		ID_HEX=`echo ${line} | grep '^Stream' | cut -d '[' -f 2 | cut -d ']' -f 1`
+		if [[ ! $ID_HEX == "" ]]; then
+			echo "ID $(($ID_HEX)) - `echo ${line} | grep ^Stream | cut -d ' ' -f 3-`"
+		fi
+	done  < ${1}
+}
+
+	## 'decode' lsdvd output file
+lsdvd-decode(){
+	while read line
+	do
+		ID_HEX=`echo ${line} | grep -vE '^Title|VTS' | grep -o '.\{5\}$' | sed s/,//`
+		if [[ ! $ID_HEX == "" ]]; then
+			echo "ID $(($ID_HEX)) - `echo ${line} | sed s/', Stream id: ....'//`"
+		fi
+	done  < ${1}
+}
+
+## pre-rip source choice
 
 case $SOURCE in
 BD )
-	if [ ! -f $M2TS_FILE ]
+	if [ ! -f "$M2TS_FILE" ]
 	then
 		echo -ne "\n *************************************\n"
 		echo " M2TS_FILE $M2TS_FILE does not exists ! Exiting..."
@@ -38,38 +69,51 @@ BD )
 		exit 1
 	fi
 
-	echo $M2TS_FILE
-
-	echo -ne "\n *************************************\n"
+	echo -ne " *************************************\n"
 	echo " MPlayer informations"
-	echo -ne " *************************************\n"
-	mplayer -v -vo null -ao null -frames 0 -identify "$M2TS_FILE" 2>/dev/null | grep -E "^ID_*|^VIDEO*|^AUDIO*|^Selected|^==="
-	
-	echo -ne "\n *************************************\n"
-	echo " FFmpeg informations"
-	echo -ne " *************************************\n"
-	# This program is not developed anymore and is only provided for compatibility. Use avconv instead (see Changelog for the list of incompatible changes).
-	ffmpeg -i $M2TS_FILE -loglevel 0 2>&1| grep -E "Duration|Stream"
+	mplayer -v -vo null -ao null -frames 0 -identify "$M2TS_FILE" 2>/dev/null > M2TS-mplayer.info
+	echo "-> see M2TS-mplayer.info"
 
-	echo -ne "\n *************************************\n"
-	echo " Crop detection"
 	echo -ne " *************************************\n"
+	echo " avconv informations"
+	avconv -i "$M2TS_FILE" 2> M2TS-avconv.info
+	echo "-> see M2TS-avconv.info"
+
+	echo -ne " *************************************\n"
+	echo M2TS_FILE=\"$M2TS_FILE\"
 	cropdetect $M2TS_FILE
+
+	echo -ne "*************************************\n"
+	echo "All tracks:"
+	avconv-decode M2TS-avconv.info
 	;;
 DVD )
 	check_lsdvd
-	echo -ne "\n *************************************\n"
-	echo " lsdvd"
-	echo -ne " *************************************\n"
-	lsdvd -acsv /dev/dvd | grep -v xx
 
-	echo -ne "\n *************************************\n"
+	echo -ne "*************************************\n"
+	echo " lsdvd"
+ 	lsdvd -acsv /dev/dvd 2>/dev/null > DVD-lsdvd.info
+	echo "-> see DVD-lsdvd.info"
+
+	echo -ne "*************************************\n"
 	echo " MPlayer"
-	echo -ne " *************************************\n"
-	mplayer -osdlevel 3 -ss $VIDEO_BEGIN -ao null dvd://$DVD_TITLE_NUMBER -chapter $CHAPTERS -vf cropdetect
-        ;;
+	mplayer -v -vo null -ao null -frames 0 -identify /dev/dvd 2>/dev/null > DVD-mplayer.info
+	echo "-> see DVD-player.info"
+
+	echo -ne "*************************************\n"
+	grep "^Disc Title" DVD-lsdvd.info
+	DVD_TITLE_NUMBER=`grep "^Longest track:" DVD-lsdvd.info | sed s/'Longest track: '//`
+	echo "Longest track chosen for crop detection: Title $DVD_TITLE_NUMBER"
+	cropdetect "dvd://$DVD_TITLE_NUMBER"
+	lsdvd -t $DVD_TITLE_NUMBER -acsv /dev/dvd 2>/dev/null | grep -E "Subtitle|Audio|VTS" > DVD-lsdvd-T$DVD_TITLE_NUMBER.info
+	grep VTS DVD-lsdvd-T$DVD_TITLE_NUMBER.info
+
+	echo -ne "*************************************\n"
+	echo "Audio/Subtitles tracks (Title $DVD_TITLE_NUMBER):"
+	lsdvd-decode DVD-lsdvd-T$DVD_TITLE_NUMBER.info
+	;;
 ISO )
-	if [ ! -f $ISO_FILE ]
+	if [ ! -f "$ISO_FILE" ]
 	then
 		echo -ne "\n *************************************\n"
 		echo " ISO_FILE $ISO_FILE does not exists ! Exiting..."
@@ -78,25 +122,31 @@ ISO )
 	fi
 
 	check_lsdvd
-	echo -ne "\n *************************************\n"
+	echo -ne "*************************************\n"
 	echo " lsdvd"
-	echo -ne " *************************************\n"
-	lsdvd -acsv $ISO_FILE 2>/dev/null | grep -v xx
+	lsdvd -acsv "$ISO_FILE" 2>/dev/null > ISO-lsdvd.info
+	echo "-> see ISO-lsdvd.info"
 	
 	echo -ne "*************************************\n"
 	echo " MPlayer informations"
-	echo -ne " *************************************\n"
-	mplayer -v -vo null -ao null -frames 0 -identify "$ISO_FILE" 2>/dev/null | grep -E "^ID_*|^VIDEO*|^AUDIO*|^Selected|^==="
+	mplayer -v -vo null -ao null -frames 0 -identify "$ISO_FILE" 2>/dev/null > ISO-mplayer.info
+	echo "-> see ISO-lsdvd.info"
 
-	echo -ne "\n *************************************\n"
-	echo " FFmpeg informations"
-	echo -ne " *************************************\n"
-	ffmpeg -i $ISO_FILE -loglevel 0 2>&1 | grep -E "Duration|Stream"
+	echo -ne "*************************************\n"
+	echo " avconv informations"
+	avconv -i "$ISO_FILE" 2> ISO-avconv.info
+	echo "-> see ISO-avconv.info"
 
-	echo -ne "\n *************************************\n"
-	echo " Crop detection: Title $DVD_TITLE_NUMBER"
-	echo -ne " *************************************\n"
+	echo -ne "*************************************\n"
+	echo ISO_FILE=\"$ISO_FILE\"
+	grep "^Disc Title" ISO-lsdvd.info
+	DVD_TITLE_NUMBER=`grep "^Longest track:" ISO-lsdvd.info | sed s/'Longest track: '//`
+	echo "Longest track chosen for crop detection: Title $DVD_TITLE_NUMBER"
 	cropdetect "-dvd-device $ISO_FILE dvd://$DVD_TITLE_NUMBER"
+
+	echo -ne "*************************************\n"
+	echo "Audio/Subtitles tracks:"
+	avconv-decode ISO-avconv.info
 	;;
 * )
         echo -ne "\n Media source problem: make sure you filled DVD ISO BD in configuration file\n"
